% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding.R
\docType{data}
\name{coding}
\alias{coding}
\title{coding Data Set}
\format{
A data frame with 20 rows and 2 columns
\describe{
    \item{DAY}{days elapsed}
    \item{ERRORS}{numbers of program modules with errors, out of 25}
}
}
\usage{
coding
}
\description{
A data set containing instances of days elapsed and the number of program
modules with errors, from 25 randomly-selected program modules
}
\keyword{datasets}
