% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire Data Set}
\format{
A data frame with 15 rows and 2 columns
\describe{
    \item{DISTANCE}{distance from fire station in miles}
    \item{DAMAGE}{x1000 dollars in damage}
}
}
\usage{
fire
}
\description{
A data set containing instances of fires, alongside data about its distance
from the fire station and the cost of damages
}
\keyword{datasets}
