% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltMean.R
\name{mycltMean}
\alias{mycltMean}
\title{Generates an n-sized random sample from the uniform distribution for a given
amount of iterations. Then, calculates the mean value for each simulated
iterations to form a histogram of this data.}
\usage{
mycltMean(n, iter)
}
\arguments{
\item{n}{the size of the random sample}

\item{iter}{the number of iterations within the simulation}
}
\value{
the vector containing the vector of mean values for each iteration's
random sample
}
\description{
Generates an n-sized random sample from the uniform distribution for a given
amount of iterations. Then, calculates the mean value for each simulated
iterations to form a histogram of this data.
}
