% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{Takes in the DDT data set and filters the data to only contain a given SPECIES.
Next, it saves this filtered data to a CSV file. Then, it graphs LENGTH vs
WEIGHT, colored according to RIVER. Afterwards, it creates a relative
frequency table, in relation to the RIVER variable. Finally, it outputs
the DDT data set before filtering, the data set after filtering, and
the relative frequency table from before.}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{the data frame for the data}

\item{SPECIES}{the SPECIES type to use to filter}
}
\value{
the list of the data before filtering, the data after filtering,
and the relative frequency table for RIVER.
}
\description{
Takes in the DDT data set and filters the data to only contain a given SPECIES.
Next, it saves this filtered data to a CSV file. Then, it graphs LENGTH vs
WEIGHT, colored according to RIVER. Afterwards, it creates a relative
frequency table, in relation to the RIVER variable. Finally, it outputs
the DDT data set before filtering, the data set after filtering, and
the relative frequency table from before.
}
