% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Calculate the value of y = beta_0 + beta_1 \* x + beta_2 \* I (x > x_k)
I (x > x_k) is 1 when x > x_k and 0 otherwise.
x_k = 18}
\usage{
myf(x, coef)
}
\arguments{
\item{x}{a number}

\item{coef}{array that holds the coefficients for our function (beta_0, beta_1, beta_2)}
}
\value{
y = beta_0 + beta_1 \* x + beta_2 \* I (x > x_k)
}
\description{
Calculate the value of y = beta_0 + beta_1 \* x + beta_2 \* I (x > x_k)
I (x > x_k) is 1 when x > x_k and 0 otherwise.
x_k = 18
}
