% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Given data, a probability-based function, and a range of potential values,
this function determines the potential value that has the highest likelihood
of generating this data.}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function applied to the range of potential values}

\item{x}{Data, used to determine the maximum likelihood outcome}

\item{theta}{Range of values, with the value with the highest likelihood
being chosen}
}
\value{
List, containing the index of the maximum likelihood outcome, the
maximum likelihood outcome, the maximum likelihood, and the slope of
the maximum likelihood outcome
}
\description{
Given data, a probability-based function, and a range of potential values,
this function determines the potential value that has the highest likelihood
of generating this data.
}
