% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Given a probability-based function and a range of potential values,
determines the value of theta that has the highest likelihood of occurring.}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{Function applied to the range of potential values (theta)}

\item{theta}{Range of values, with the value with the highest likelihood
being chosen}
}
\value{
theta value, with the maximum likelihood
}
\description{
Given a probability-based function and a range of potential values,
determines the value of theta that has the highest likelihood of occurring.
}
