% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Displays a normal curve (given mu and sigma), shades in the area between
the curve and the x-axis from -Inf (approximated to 3 SD's away from the mean)
to x=a (given a), and returns the area as a form of probability.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the normal distribution}

\item{sigma}{the standard deviation of the normal distribution}

\item{a}{the right-most value, such that the probability is calculated from
negative infinity to x=a}
}
\value{
the probability of P(Y <= a)
}
\description{
Displays a normal curve (given mu and sigma), shades in the area between
the curve and the x-axis from -Inf (approximated to 3 SD's away from the mean)
to x=a (given a), and returns the area as a form of probability.
}
