% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Generates a given amount of iterations for a given amount of random samples
between 1 and 10 (inclusive) with replacement. Then, graphs the data
to a barplot, before waiting to start the next iteration.}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{the amount of times a number from [1:10] will be selected by random
sampling in a given iteration}

\item{iter}{the amount of experiments that this random sampling will occur}

\item{time}{the time between any two iterations}
}
\description{
Generates a given amount of iterations for a given amount of random samples
between 1 and 10 (inclusive) with replacement. Then, graphs the data
to a barplot, before waiting to start the next iteration.
}
